# ENH: Speed up common/non_neighbors by using _adj dict operations