I'm not sure why we should not use direct set operations on the `_adj` dictionary to find common neighbors. This will require a deprecation probably as we will change the behavior of the return type (`generator` to `set`).

This was spurred by https://github.com/networkx/networkx/discussions/7243.