@dschult 2024-01-21
This looks good. The reasons for the generator approach are old and no longer relevant (sets were slower and _adj was only used in the base classes).

I am not sure that we need a deprecation -- it is true that the type of the return value is changed from generator to set. But that will only break code that is calling `next` directly on the return value. Anything using `len` or looping over the result will work quite nicely. Searching github for nx.common_neighbors shows no cases where it won't just work (of the first 3 pages of results -- I can look further). But I tend to lean toward "just putting it in", so I'm biased here. :)  Are there cases where a set would not work but in a silent way?

There are quite a few other functions in the `function.py` module that might be similarly sped up by large factors. That code has been looked at since the sets have been improved and it's pretty clear that performance wasn't the intent in some choices (like using `neighbors(G, u, v)` instead of `G.neighbors(u, v)` in `nonneighbors`).