@dschult 2024-01-22
The non_neighbors looks good too.  
Do you want to change all the `G.adj[` and `G[` to `G._adj[` in this module? Or am I trying to put too much into this PR... :}
Thanks!