@MridulS 2024-01-22
I think I have covered the all `adj` calls in functions.py, not sure if I missed something!