@dschult 2024-01-22
> This changes behavior. Graph.__contains__ has a try/except statement,

Nice catch!  

I guess that raises the question of whether a single try/except with a loop inside is better than many try/excepts inside a loop. That should be a python dependent question maybe answered elsewhere -- but I didn't find in a quick search.