@rossbar 2024-01-25
I hope this doesn't distract the discussion too much... but since this PR largely deals with performance, I thought it'd be a good opportunity to try adding some benchmarks!

I went ahead and pushed up benchmarks for `nx.non_neighbors` (bcd2d2e) and `nx.common_neighbors` (6a36798) just so others can try out benchmarks while reviewing. You can then use these new benchmarks to compare the performance improvements of these changes relative to `main`. For example:

```
# --bench accepts a regex for limiting which benchmarks are run
asv continuous --bench CommonNeighbors main speedup_neighbors
```

Which on my system (m1 mac running asahi linux) gives

<details>
  <Summary><code>asv continuous --bench CommonNeighbors main speedup_neighbors</code></summary>
<pre>
| Change   | Before [d6569adf] &lt;main&gt;   | After [6a367983] &lt;speedup_neighbors&gt;   |   Ratio | Benchmark (Parameter)                                          |
|----------|----------------------------|----------------------------------------|---------|----------------------------------------------------------------|
| -        | 1.18±0.02μs                | 742±10ns                               |    0.63 | benchmark_neighbors.CommonNeighbors.time_star_rim_rim(100)     |
| -        | 1.18±0.03μs                | 735±10ns                               |    0.62 | benchmark_neighbors.CommonNeighbors.time_star_rim_rim(1000)    |
| -        | 1.19±0.01μs                | 728±7ns                                |    0.61 | benchmark_neighbors.CommonNeighbors.time_star_rim_rim(10)      |
| -        | 3.79±0.1μs                 | 941±10ns                               |    0.25 | benchmark_neighbors.CommonNeighbors.time_complete(10)          |
| -        | 5.07±0.08μs                | 665±5ns                                |    0.13 | benchmark_neighbors.CommonNeighbors.time_star_center_rim(10)   |
| -        | 30.5±0.3μs                 | 3.80±0.05μs                            |    0.12 | benchmark_neighbors.CommonNeighbors.time_complete(100)         |
| -        | 311±8μs                    | 28.6±0.4μs                             |    0.09 | benchmark_neighbors.CommonNeighbors.time_complete(1000)        |
| -        | 44.0±0.6μs                 | 683±10ns                               |    0.02 | benchmark_neighbors.CommonNeighbors.time_star_center_rim(100)  |
| -        | 439±7μs                    | 732±10ns                               |    0    | benchmark_neighbors.CommonNeighbors.time_star_center_rim(1000) 
</pre>
</details>

Anyways - IMO it's really nice to be able to get a sense for the magnitude of performance changes in these types of PRs[^1], and I wanted to share my experimentation! I'm also happy to pull these two commits back out and put them in a separate PR!

[^1]: These changes are actually a little trickier to benchmark than you'd expect since the return types are switching from a generator to a set, so the benchmarks have to have something baked-in to consume the generator and ensure the outputs are comparable.